Call window: Ellipsize alias and status
If the user alias is too long both the hangup and the dialpad buttons
get pushed into a drop-down menu. Even worse when the alias doesn't fit,
it just isn't shown at all anymore.

Given the username is already printed in the title and the user tends to
know who they're calling with, ellipsizing the alias isn't an issue.
Which has the nice side-effect of keeping the hangup and dialpad buttons
on the toolbar in all cases like they should be.

Unfortunately as pango doesn't ellipsize if it still has enough height
left to put more lines in. So put the alias and status in seperate
labels in a vbox which isn't set to fill.