Call: Let the floating toolbar be above the preview
The call window has previews in the four corners, before the lower
ones were packed in a way that they would appear above the floating
toolbar which looks strange when the floating toolbar disappears.

Instead allow the lower preview to be below the floating toolbar so
they're evenly spread out.