audio input: Switch to stream volumes
Newer gstreamer & pulseaudio support the stream volume interface for
input as well. Prefer this over using the mixer interface as it's
both simpler and actually does as intended. Besides that the
mixer interface is buggy and seems to not correctly adjust the
current input device if the source was switch to a non-default input.

As an extra put in a volume element to locally enforce the current mute
state. This ensure that whatever happens, if the UI says mute, the
stream is guaranteed to be muted. This prevents awkward situations if
the source element doesn't support stream volumes or the notification
is buggy (like with current pulsesrc in releases).