roster-view: add EmpathyRosterGroup and sort contacts accordingly
The GtkExpander doesn't contain the contacts as its children because the view
needs to be have full control of which rows as displayed (for live search for
example). So instead we trick the view sort function to display the contact
associated with the group at the right position.

Also, we need to keep the structure flat to not break keyboard navigation
inside the widget.