Rely on the empathy factory rather than 're-preparing' the AM
tp_account_manager_prepare_all_async() seems to be a bad idea so, instead,
let's just rely the empathy factory and set it as default.
This makes the code much more symetric with empathy-accounts and shouldn't
prepare that much useless features as we are not going to prepare the contact
list anyway.